perf: upload wide, cast on host — 721.6 -> 586.6 ms/frame
Two transfer fixes, no accuracy change, trajectory still identical:

  cast to bf16 on the host before from_torch   721.6 -> 600.8 ms
  upload the grid wide, reshape on device      600.8 -> 586.6 ms

from_torch costs per row rather than per byte. The same 5.9 MB moves at
50 MB/s shaped [C, N, 1, 2] and 4516 MB/s shaped [C, N*2/256, 256] -- two
million rows against twelve thousand. The sampling grid has the worst possible
shape for this and was 161 ms of a 499 ms DFA call. Uploading wide and
reshaping on device is 118.0 -> 7.7 ms, bit-identical output.

That corrects an earlier conclusion: assembling the grid on device was rejected
for measuring slower, blamed on tile padding. The rejection stands, the reason
was wrong, and the real cause was fixable without moving the assembly.

Also measured, for what comes next: grid_sample's cost is per point and
independent of channel count -- 32 and 256 channels take the same time -- so it
is the coordinate work that is expensive, not the gather. That is what
grid_precompute addresses, and it means grid_compact buys only the
visible-point fraction of the per-point cost.

+++ model/sparsedrive.py
@@ -48,7 +48,8 @@ class TtSparseDrive:
     def features(self, imgs):
         """imgs [cams, 3, H, W] -> per-level ttnn NHWC + the image tokens."""
         x = torch.nn.functional.pad(imgs.permute(0, 2, 3, 1), (0, 1))
-        xt = ttnn.from_torch(x.reshape(1, 1, self.C * self.H * self.W, 4).contiguous(),
+        xt = ttnn.from_torch(
+            x.reshape(1, 1, self.C * self.H * self.W, 4).bfloat16().contiguous(),
                              dtype=ttnn.bfloat16, layout=ttnn.ROW_MAJOR_LAYOUT,
                              device=self.dev, mesh_mapper=_mapper(self.dev)[1])
         outs = self.fpn(self.backbone(xt))
@@ -60,7 +61,7 @@ class TtSparseDrive:
             v = to_host(t, self.dev).float().reshape(-1, c)[:self.C * h * w]
             v = v.reshape(self.C, h, w, c)
             levels.append(ttnn.from_torch(
-                v.contiguous(), layout=ttnn.ROW_MAJOR_LAYOUT, device=self.dev,
+                v.bfloat16().contiguous(), layout=ttnn.ROW_MAJOR_LAYOUT, device=self.dev,
                 dtype=ttnn.bfloat16, mesh_mapper=_mapper(self.dev)[1]))
             last = v
         # image tokens for the velocity branch's cross-attention: the coarsest
